Win64 fixes
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@6413 212acab6-be3b-0410-9dea-997c60f758d6

+++ Contrib/NSISdl/nsisdl.cpp
@@ -30,7 +30,7 @@
 
 #include <nsis/pluginapi.h> // nsis plugin
 
-void *operator new( unsigned int num_bytes )
+void *operator new( size_t num_bytes )
 {
   return GlobalAlloc(GPTR,num_bytes);
 }
